LayoutTests/http/wpt/service-workers/service-worker-no-install-event.html

<!doctype html><!-- webkit-test-runner [ ServiceWorkerInstallEventEnabled=false ] -->
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/routines.js"></script>
</head>
<body>
<script>
promise_test(async (test) => {
    const registration = await navigator.serviceWorker.register("service-worker-no-install-event.js");
    const worker = registration.installing;

    await waitForState(registration.installing, "activated");

    worker.postMessage("test");
    assert_equals(await new Promise(resolve => navigator.serviceWorker.onmessage = e => resolve(e.data)), "No InstallEvent - no addRoutes");
}, "add route when not installing");
</script>
</body>
</html>
